#105511

Related work items: #105511

+++ Logitude/Unifreight.Data/AmitalModel/AmitalContext.Template.cs
@@ -12233,13 +12233,13 @@ namespace Unifreight.Data.AmitalModel
                     .HasColumnName(@"FILE_NO")
                     .IsRequired()
                     .HasDatabaseGeneratedOption(System.ComponentModel.DataAnnotations.Schema.DatabaseGeneratedOption.None)
-                    .HasColumnType("int");
+                    .HasColumnType("bigint");
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.LINENO)
                     .HasColumnName(@"LINE_NO")
                     .IsRequired()
                     .HasDatabaseGeneratedOption(System.ComponentModel.DataAnnotations.Schema.DatabaseGeneratedOption.None)
-                    .HasColumnType("int");
+                    .HasColumnType("decimal");
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.PRATMEHES)
                     .HasColumnName(@"PRAT_MEHES")
@@ -12248,7 +12248,7 @@ namespace Unifreight.Data.AmitalModel
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.GOODSNO)
                     .HasColumnName(@"GOODS_NO")
-                    .HasColumnType("int");
+                    .HasColumnType("decimal");
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.TAXTYPE)
                     .HasColumnName(@"TAX_TYPE")
@@ -12293,7 +12293,7 @@ namespace Unifreight.Data.AmitalModel
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.ADDIMPORT)
                     .HasColumnName(@"ADD_IMPORT")
-                    .HasColumnType("float");
+                    .HasColumnType("decimal");
             modelBuilder.Entity<CCUTAX>()
                 .Property(p => p.PRATMEHESN)
                     .HasColumnName(@"PRAT_MEHES_N")
